+++ build.gradle
@@ -1,8 +1,8 @@
 buildscript {
     ext {
-        kotlin_version = '1.5.31'
+        kotlin_version = '1.6.10'
         hilt_version = '2.40'
-        compose_version = '1.0.5'
+        compose_version = '1.1.0'
     }
     repositories {
         google()
